+++ src/schema.ts
@@ -15,11 +15,13 @@ import {
   Operator,
 } from './operators'
 import {
+  OrderByItemInfo,
   DIRECTION_NOT_EXIST,
   ASC,
   DESC,
-  OrderByItemInfo,
-  OrderByNullsPosition,
+  NULLS_POSITION_NOT_EXIST,
+  NULLS_FIRST,
+  NULLS_LAST,
 } from './orderBy'
 import { SelectItemInfo } from './select'
 import { escapeDoubleQuote } from './util'
@@ -106,35 +108,35 @@ export abstract class Column {
   }
 
   public get asc(): OrderByItemInfo {
-    return new OrderByItemInfo(this, ASC, OrderByNullsPosition.NOT_EXIST)
+    return new OrderByItemInfo(this, ASC, NULLS_POSITION_NOT_EXIST)
   }
 
   public get desc(): OrderByItemInfo {
-    return new OrderByItemInfo(this, DESC, OrderByNullsPosition.NOT_EXIST)
+    return new OrderByItemInfo(this, DESC, NULLS_POSITION_NOT_EXIST)
   }
 
   public get nullsFirst(): OrderByItemInfo {
-    return new OrderByItemInfo(this, DIRECTION_NOT_EXIST, OrderByNullsPosition.NULLS_FIRST)
+    return new OrderByItemInfo(this, DIRECTION_NOT_EXIST, NULLS_FIRST)
   }
 
   public get nullsLast(): OrderByItemInfo {
-    return new OrderByItemInfo(this, DIRECTION_NOT_EXIST, OrderByNullsPosition.NULLS_LAST)
+    return new OrderByItemInfo(this, DIRECTION_NOT_EXIST, NULLS_LAST)
   }
 
   public get ascNullsFirst(): OrderByItemInfo {
-    return new OrderByItemInfo(this, ASC, OrderByNullsPosition.NULLS_FIRST)
+    return new OrderByItemInfo(this, ASC, NULLS_FIRST)
   }
 
   public get descNullsFirst(): OrderByItemInfo {
-    return new OrderByItemInfo(this, DESC, OrderByNullsPosition.NULLS_FIRST)
+    return new OrderByItemInfo(this, DESC, NULLS_FIRST)
   }
 
   public get ascNullsLast(): OrderByItemInfo {
-    return new OrderByItemInfo(this, ASC, OrderByNullsPosition.NULLS_LAST)
+    return new OrderByItemInfo(this, ASC, NULLS_LAST)
   }
 
   public get descNullsLast(): OrderByItemInfo {
-    return new OrderByItemInfo(this, DESC, OrderByNullsPosition.NULLS_LAST)
+    return new OrderByItemInfo(this, DESC, NULLS_LAST)
   }
 
   public toString() {